+++ topobank/taskapp/serializers.py
@@ -11,8 +11,8 @@ class TaskStateModelSerializer(serializers.HyperlinkedModelSerializer):
         abstract = True
         model = TaskStateModel
         fields = [
-            "duration",
             "error",
+            "task_duration",
             "task_progress",
             "task_state",
             "task_memory",
@@ -22,8 +22,8 @@ class TaskStateModelSerializer(serializers.HyperlinkedModelSerializer):
 
     default_error_messages = {"read_only": "This field is read only"}
 
-    duration = serializers.DurationField(read_only=True)
     error = serializers.CharField(source="get_task_error", read_only=True)
+    task_duration = serializers.DurationField(read_only=True)
     task_progress = serializers.FloatField(source="get_task_progress", read_only=True)
     task_state = serializers.CharField(source="get_task_state", read_only=True)
     celery_task_state = serializers.CharField(source="get_celery_state", read_only=True)