+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog for *TopoBank*
 
+# 1.61.0 (not yet released)
+
+- ENH: Import properties from ZIP container
+- MAINT: Renamed `import_surfaces` to `import_datasets`
+- MAINT: Added `task_submission_time`, which allows better failure detection
+- MAINT: Renamed `start_time` to `task_start_time` and `end_time` to `task_end_time`
+
+# 1.60.1 (2025-04-27)
+
+- BUG: More robust task state reporting
+
 # 1.60.0 (2025-04-27)
 
 - API: Asynchronous preparation of ZIP downloads (#249)